# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponseDto" }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Bootstrap and return the authenticated account profile",
        "tags": ["profiles"]
      },
      "patch": {
        "operationId": "updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMeDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponseDto" }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Update the authenticated account public identity and privacy settings",
        "tags": ["profiles"]
      }
    },
...
    "/v1/webhooks/hyperwallet": {
      "post": {
        "operationId": "receive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HyperwalletWebhookDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HyperwalletWebhookAcceptedDto"
                }
              }
            }
          }
        },
        "security": [{ "hyperwallet-webhook": [] }],
        "summary": "Durably accept a Hyperwallet notification for reconciliation",
        "tags": ["provider webhooks"]
      }
    },
    "/v1/operator/payout-release-control": {
      "get": {
        "operationId": "get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutReleaseControlResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Get the global payout release safety control",
        "tags": ["operator payouts"]
      }
    },
    "/v1/operator/payout-release-control/status-action": {
      "post": {
        "operationId": "change",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayoutReleaseControlActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutReleaseControlResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Pause or resume new payout release reservations",
        "tags": ["operator payouts"]
      }
    },
    "/v1/operator/payout-claims/{claimId}/review": {
      "get": {
        "operationId": "getReview",
        "parameters": [
          {
            "name": "claimId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorPayoutClaimReviewDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Get a privacy-safe authoritative payout claim review",
        "tags": ["operator payouts"]
      }
    },
    "/v1/operator/payout-claims/{claimId}/approve": {
      "post": {
        "operationId": "approve",
        "parameters": [
          {
            "name": "claimId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorPayoutActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorPayoutClaimResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Approve a reviewed winner claim for payee setup",
        "tags": ["operator payouts"]
      }
    },
    "/v1/operator/payout-claims/{claimId}/release": {
      "post": {
        "operationId": "release",
        "parameters": [
          {
            "name": "claimId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorPayoutActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorPayoutClaimResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Release a ready claim to Hyperwallet exactly once",
        "tags": ["operator payouts"]
      }
    },
    "/v1/payout-claims/me": {
      "get": {
        "operationId": "getMyClaim",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutClaimResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Return the current approved winner payout claim",
        "tags": ["payouts"]
      }
    },
    "/v1/payout-claims/{claimId}/portal-action": {
      "post": {
        "operationId": "openPortal",
        "parameters": [
          {
            "name": "claimId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Provision hosted payee setup and return the approved portal URL",
        "tags": ["payouts"]
      }
    },
...
  },
  "info": {
    "title": "GoGymGo API",
    "description": "Authoritative API for GoGymGo accounts, verified competition activity, ledgers, results, and payouts.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "firebase": {
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "type": "http",
        "description": "Firebase ID token verified by the GoGymGo API."
      },
      "hyperwallet-webhook": {
        "type": "http",
        "scheme": "basic",
        "description": "Dedicated credentials configured only for Hyperwallet."
      }
    },
    "schemas": {
...
      "PrivacySettingsDto": {
        "type": "object",
        "properties": {
          "showRegion": { "type": "boolean" },
          "showStats": { "type": "boolean" }
        },
        "required": ["showRegion", "showStats"]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "callsign": { "type": "string" },
          "email": { "type": "string", "nullable": true },
          "emailVerified": { "type": "boolean" },
          "publicIdentityMode": {
            "type": "string",
            "enum": ["private", "alias", "real_name"]
          },
          "publicName": { "type": "string", "nullable": true },
          "privacySettings": {
            "$ref": "#/components/schemas/PrivacySettingsDto"
          },
          "roles": { "type": "array", "items": { "type": "string" } },
          "status": {
            "type": "string",
            "enum": ["active", "suspended", "deleted"]
          },
          "version": { "type": "number" }
        },
        "required": [
          "id",
          "callsign",
          "email",
          "emailVerified",
          "publicIdentityMode",
          "publicName",
          "privacySettings",
          "roles",
          "status",
          "version"
        ]
      },
      "UpdatePrivacySettingsDto": {
        "type": "object",
        "properties": {
          "showRegion": { "type": "boolean" },
          "showStats": { "type": "boolean" }
        }
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "publicIdentityMode": {
            "type": "string",
            "enum": ["private", "alias", "real_name"]
          },
          "publicName": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "nullable": true
          },
          "privacySettings": {
            "$ref": "#/components/schemas/UpdatePrivacySettingsDto"
          }
        }
      },
...
      "HyperwalletWebhookDto": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "maxLength": 64 },
          "type": { "type": "string", "maxLength": 160 },
          "createdOn": { "type": "string", "format": "date-time" },
          "data": { "type": "object" }
        },
        "required": ["token", "type"]
      },
      "HyperwalletWebhookAcceptedDto": {
        "type": "object",
        "properties": { "accepted": { "type": "boolean" } },
        "required": ["accepted"]
      },
      "PayoutReleaseControlResponseDto": {
        "type": "object",
        "properties": {
          "paused": { "type": "boolean" },
          "reason": { "type": "string" },
          "version": { "type": "number", "minimum": 1 }
        },
        "required": ["paused", "reason", "version"]
      },
      "PayoutReleaseControlActionDto": {
        "type": "object",
        "properties": {
          "expectedVersion": { "type": "number", "minimum": 1 },
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "action": { "type": "string", "enum": ["pause", "resume"] }
        },
        "required": ["expectedVersion", "reason", "action"]
      },
      "OperatorPayoutParticipantReviewDto": {
        "type": "object",
        "properties": {
          "accountStatus": {
            "type": "string",
            "enum": ["active", "deleted", "suspended"]
          },
          "eligible": { "type": "boolean" },
          "emailPresent": { "type": "boolean" },
          "emailVerified": { "type": "boolean" }
        },
        "required": [
          "accountStatus",
          "eligible",
          "emailPresent",
          "emailVerified"
        ]
      },
      "OperatorPayoutAccountReviewDto": {
        "type": "object",
        "properties": {
          "provisioned": { "type": "boolean" },
          "ready": { "type": "boolean" },
          "payeeStatus": { "type": "string", "nullable": true }
        },
        "required": ["provisioned", "ready", "payeeStatus"]
      },
      "OperatorPayoutPaymentReviewDto": {
        "type": "object",
        "properties": { "status": { "type": "string", "nullable": true } },
        "required": ["status"]
      },
      "OperatorPayoutClaimReviewDto": {
        "type": "object",
        "properties": {
          "amountMinor": { "type": "number" },
          "approvedAt": { "type": "string", "nullable": true },
          "competitionId": { "type": "string", "format": "uuid" },
          "competitionName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "currency": { "type": "string", "enum": ["CAD", "MXN", "USD"] },
          "drawId": { "type": "string", "format": "uuid" },
          "failureCode": { "type": "string", "nullable": true },
          "id": { "type": "string", "format": "uuid" },
          "participant": {
            "$ref": "#/components/schemas/OperatorPayoutParticipantReviewDto"
          },
          "paidAt": { "type": "string", "nullable": true },
          "payoutAccount": {
            "$ref": "#/components/schemas/OperatorPayoutAccountReviewDto"
          },
          "payoutRank": { "type": "number", "minimum": 1 },
          "payment": {
            "$ref": "#/components/schemas/OperatorPayoutPaymentReviewDto"
          },
          "releaseControl": {
            "$ref": "#/components/schemas/PayoutReleaseControlResponseDto"
          },
          "provider": { "type": "string", "enum": ["hyperwallet"] },
          "status": { "type": "string" },
          "version": { "type": "number", "minimum": 1 },
          "winnerId": { "type": "string", "format": "uuid" }
        },
        "required": [
          "amountMinor",
          "approvedAt",
          "competitionId",
          "competitionName",
          "createdAt",
          "currency",
          "drawId",
          "failureCode",
          "id",
          "participant",
          "paidAt",
          "payoutAccount",
          "payoutRank",
          "payment",
          "releaseControl",
          "provider",
          "status",
          "version",
          "winnerId"
        ]
      },
      "OperatorPayoutActionDto": {
        "type": "object",
        "properties": {
          "expectedVersion": { "type": "number", "minimum": 1 },
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 }
        },
        "required": ["expectedVersion", "reason"]
      },
      "OperatorPayoutClaimResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": [
              "pending_review",
              "action_required",
              "verification_pending",
              "ready",
              "processing",
              "paid",
              "failed",
              "cancelled"
            ]
          },
          "version": { "type": "number" }
        },
        "required": ["id", "status", "version"]
      },
      "PayoutClaimResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "amountMinor": { "type": "number" },
          "competitionLabel": { "type": "string" },
          "currency": { "type": "string", "enum": ["CAD", "MXN", "USD"] },
          "provider": { "type": "string", "enum": ["hyperwallet"] },
          "status": {
            "type": "string",
            "enum": ["action-required", "verification-pending", "ready", "paid"]
          }
        },
        "required": [
          "id",
          "amountMinor",
          "competitionLabel",
          "currency",
          "provider",
          "status"
        ]
      },
      "PortalActionResponseDto": {
        "type": "object",
        "properties": { "portalUrl": { "type": "string", "format": "uri" } },
        "required": ["portalUrl"]
      },
...
    }
  }
}